Use stdint.h types everywhere

+++ src/grid.h
@@ -1,7 +1,7 @@
 #ifndef __GRID_H__
 #define __GRID_H__
 
-void print(int *sudoku);
-void solve(int *sudoku);
+void print(uint8_t *sudoku);
+void solve(uint8_t *sudoku);
 
 #endif
